Move RelationshipProxyWrappingValue out of public API
