+++ community/cypher/cypher/src/main/scala/org/neo4j/cypher/internal/compatibility/v3_3/runtime/executionplan/procs/ProcedureExecutionResult.scala
@@ -34,9 +34,10 @@ import org.neo4j.cypher.result.QueryResult.{QueryResultVisitor, Record}
 import org.neo4j.graphdb.Notification
 import org.neo4j.graphdb.spatial.{Geometry, Point}
 import org.neo4j.kernel.impl.util.ValueUtils
+import org.neo4j.kernel.impl.util.ValueUtils._
 import org.neo4j.values.AnyValue
 import org.neo4j.values.storable.Values
-import org.neo4j.values.storable.Values.{of => _, _}
+import org.neo4j.values.storable.Values._
 
 /**
   * Execution result of a Procedure
@@ -87,7 +88,7 @@ class ProcedureExecutionResult(context: QueryContext,
   }
 
   private def transform[T](value: AnyRef, f: T => AnyValue): AnyValue = {
-    if (value == null) Values.NO_VALUE
+    if (value == null) NO_VALUE
     else f(value.asInstanceOf[T])
   }
